fingraph-project/requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.24.0
yfinance>=0.2.28
pyarrow>=12.0.0

# ML/DL
torch>=2.0.0
torch-geometric>=2.3.0
scikit-learn>=1.3.0

# API
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0

# Visualization
streamlit>=1.25.0
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utils
pyyaml>=6.0
python-dotenv>=1.0.0
joblib>=1.3.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0